requirements.txt
discord.py
line-bot-sdk==1.16.0
